requirements.txt
discord.py
python-socketio[client]
python-dotenv
aiohttp
asyncio-mqtt